refactor(upgrade-subprotocol): split into parser and handler
