+++ crates/asm/txs/upgrade/src/lib.rs
@@ -1,4 +1,4 @@
 pub mod actions;
 pub mod constants;
 pub mod error;
-pub mod parse;
+pub mod parser;
